Adding files for stable version

+++ app/src/main/java/com/project/space/CircleJoin.java
@@ -2,11 +2,11 @@ package com.project.space;
 
 public class CircleJoin {
     public String circleMemberId;
-    public double currDistance;
-    public double prevDistance;
+    public int currDistance;
+    public int prevDistance;
 
 
-    public CircleJoin(String circleMemberId,double currDistance,double prevDistance){
+    public CircleJoin(String circleMemberId,int currDistance,int prevDistance){
         this.circleMemberId = circleMemberId;
         this.currDistance = currDistance;
         this.prevDistance = prevDistance;
